Add context parameter to root KEK retrieval
Add an additional context parameter to calls to root_kek().

This context information is added to the keyblob format, so that the
relevant context can be retrieved from the keyblob and provided to
root_kek() in preparation for keyblob decryption.

For a new keyblob (under key generation, key import, or key upgrade) use
a new kek_context() trait method to ask for the current context to use.
Use this fresh context when performing root_kek() in preparation for
key encryption, and store the kek_context in the keyblob.

Bug: 197891150
Bug: 195310053
Test: VtsAidlKeyMintTargetTest
Change-Id: I4c3a0f145eab470af4239a69c5361124c0b857c5