ta: allow device HMAC key to be in hw
Add an `hmac_key_agreed` method that optionally allows the negotiated
HMAC key to be installed in hardware. If this is the case, an
implementation of a new trait is returned; this new trait encapsulates
the process of calculating an HMAC with the negotiated key.

If no trait impl is returned (as is the default), then a default
implementation of the new trait is used, which holds the HMAC key in
memory.

Bug: 197891150
Bug: 242838132
Test: VtsAidlKeyMintTargetTest
Change-Id: Ia6495b9929d5a6e4cddefef3530de7bda85e5b28